+++ crates/mimic/src/core/view.rs
@@ -1,4 +1,4 @@
-use crate::core::traits::{CreateView, TypeView, UpdateView};
+use crate::core::traits::{CreateView, FilterView, SortView, TypeView, UpdateView};
 
 // View
 pub type View<T> = <T as TypeView>::View;
@@ -8,3 +8,9 @@ pub type Create<T> = <T as CreateView>::View;
 
 // Update
 pub type Update<T> = <T as UpdateView>::View;
+
+// Filter
+pub type Filter<T> = <T as FilterView>::View;
+
+// Sort
+pub type Sort<T> = <T as SortView>::View;